pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "MCP Chat CLI with OpenAI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "mcp[cli]>=1.8.0",
    "prompt-toolkit>=3.0.51",
    "python-dotenv>=1.1.0",
]